Use slower, but more correct search (2022D21)
The brute-force search could find false-positives. I got lucky that it
did not. Changing the magic value for the initial window shows how
easily it can break. The reason is that the math function was doing
integer division. We should actually abort when a non-integer division
happens. This is now implemented (via exception throwing).

If we detect that (by catching the exception), we increase the guess
value by 1 until it no longer happens. This ensures that only actually
possible results are generated, with the drawback that the search is
going a lot slower, due to the primitive +1 increase, which can happen
a lot if we have bad luck.

This approach still feels better, though. If it runs too slow for
different inputs, we can try with different values for the start value
of window until we are lucky.