Add input.txt symlink to app dir for Kotlin
Kotlin app reads from `app/input.txt`, but CI scripts are always
downloading to the main directory, so we add a symlink to make
everything work. Symlink can safely be checked in to Git, so rule to
overwrite the usually forbidden `input.txt` is added to `.gitignore` for
the Kotlin template (and solution).