Make run.sh of (almost) all templates consistent
All run.sh scripts now support being called without arguments, in which
case they default to `input.txt`. Otherwise the first argument is used
as the filename.

Python is a bit special, as the unit test running is controlled by a
command line flag, so it is kept as it is.

Clojure and Elixir don't have support for input files besides input.txt.

Relates to #13